refactor(daemon): use kernel agent squad schema
